requirements.txt
# Database
psycopg2-binary>=2.9.0     # PostgreSQL adapter

# CSV Processing
pandas>=2.0.0              # DataFrame operations

# API (for n8n integration)
fastapi>=0.100.0           # API framework
uvicorn>=0.23.0            # ASGI server

# SFTP
paramiko>=3.0.0            # SFTP client

# Configuration
pyyaml>=6.0.0              # YAML config parsing
pydantic>=2.0.0            # Data validation

# HTTP (for webhook callbacks)
httpx>=0.25.0              # Async HTTP client

# Environment
python-dotenv>=1.0.0       # Load .env files

# Scheduling
APScheduler>=3.10.0        # Job scheduling

# Development
# pytest>=7.0.0            # Testing (uncomment when needed)
# black>=23.0.0            # Formatter (uncomment when needed)
